Unify startup code into HOST-CONSOLE, fix command line cancellation
The HOST-CONSOLE function has been receiving a LAST-RESULT value, as
well as a LAST-FAILED value...in order to print out the result of the
last execution request.

In the previous commit, HOST-CONSOLE was allowed to return either a
GROUP! or a BLOCK! to be executed, with a BLOCK! being interpreted as
an implementation sandboxing that the console was using on its own
behalf...vs. user code.  Failures when asking to execute a BLOCK!
could then be more severe, causing the console to reset the skin to
try and salvage a working REPL.  It was important to know whether the
previous request came from a BLOCK! or a GROUP!, so the LAST-FAILED
status became more complex to try and encode that information.

This changes that to just pass a third value in to each iteration of
the HOST-CONSOLE loop... *the prior code block that was requested to
have execute*.  Not only can it be examined to see if it was a BLOCK!
or a GROUP!, but other properties can be mined from it...including
directives for how to react to things like errors or halts.

Also: in the initial implementation of userspace console, the embedded
Rebol code in the host binary would run and return a HOST-START
function.  It would be the job of this function to perform startup,
and either return an INTEGER! to indicate it wanted to not run any
REPL but to terminate with that exit code...or to return a FUNCTION!
that would implement a REPL.  Then the REPL would be called in a loop.

Now the REPL has sandboxing features, allowing it to make requests of
the system to implement its own hooks in Rebol code.  And the REPL
also has a parameter indicating the last result, and the code it asked
to run which gave that result.  If the status of the code it asked
to run before is passed in as void, that can just be used to suggest
startup be run.  Then what was HOST-START can now sandbox its own
usermode code invocations (such as running strings with `--do`)

As part of this commit, an issue where scripts run on the command
line could not be canceled with Ctrl-C is resolved.

This includes a reorganization to put the "console startup" code into
the %host-console.r file, and passes a proper value (even an optional
one) to the PRINT-RESULT console skin hook vs. the molded string
of that value.