Invisible C-DEBUG-BREAK, QUOTE/SOFT, UNEVAL voids/BAR!s
The "C-DEBUG-BREAK" native is helpful for interrupting evaluation to
kick into the C debugger at a "just-in-time" moment where you want to
step through a specific call.  But if you wrote something like:

    print c-debug-break mold foo

...it took no arguments, and returned nothing.  So it would not be
possible to continue after this operation...as PRINT cannot print void.

The same mechanism that PROBE uses can't be used here.  PROBE takes an
evaluated parameter, and then passes it through as its result.  But if
C-DEBUG-BREAK did that, it would undermine the point: e.g. in the
example above, it's supposed to break into the debugger *before* the
MOLD, not after it.

This commit makes C-DEBUG-BREAK an arity-1 hard quoted operation.  It
then queues up its break, and proceeds to take the hard quoted item
and put it into the frame's cell to retrigger through the same
mechanic used by EVAL.  The effect is to appear "invisible" like PROBE,
*but* breaking before the "argument" gets evaluated.  The variations
that take actual tick counts are then broken into a separate native,
C-DEBUG-BREAK-AT, since the EVAL nature would wind up consuming the
refinement arguments.

In order to generalize this, it needs to be able to accept <end> (e.g.
so that plain C-DEBUG-BREAK entered at a prompt would not be an error).
Yet taking voids had been prohibited by hard quotes, based on the
premise that "it could never happen".  It was checked at FUNCTION!
creation time.

This prescriptivism was designed to help user understanding, since it
seeemed to suggest an ignorance of what `QUOTE ()` meant if someone
tried to make such an operator take void.  But developments on the
libRebol API has shown the need for the general principle to be
relaxed, as rebDo() API needs to be able to pass "pre-evaluated" voids
coming from C into <opt> parameters.

Therefore the check mechanically prohibiting <opt> or <end> on hard
quoted function args is scrapped in this commit.  However, QUOTE
remains unchanged--as it is an inappropriate generalized operator to be
using for splicing unevaluated slots (since it prohibits passing BAR!
for safety).

The missing operation of a "raw" hard quote is instead added as a new
function: UNEVAL.  It will blindly pass through anything--voids, BAR!s,
etc.

(Note: An important recent place that needed this operation is the
recently updated console code...which previously could not print out a
BAR! due to its use of QUOTE to suppress function evaluation.)

This motivated the need to bump the TG_Do_Count on EVAL operations,
which required splitting out the idea of starting a new expression
(which EVAL should not do) with the idea of bumping the count for
debug purposes.  It also motivated touching up the Dump_Frame_Location
code to give better results after turning the 'current' cell pointer
into a local variable in Do_Core() to coalesce eval with "looking ahead
for lookback quoting operations".

Also adds a /SOFT refinement to QUOTE, which efficiently splices the
argument into the evaluation stream vs. starting a new frame.

Additionally this renamed C-BREAK-DEBUG to C-DEBUG-BREAK, because
debug_break() is the name of the C function (picked by a third party,
not chosen by us) and I always got confused. :P