Take tick breakpoint on command line (--breakpoint NNN)
This commit uses a clever mathematical trick to round up the tick
count to a consistent point, regardless of whether some amount of
additional processing is required in usermode to process a breakpoint
request or not.  The premise and method is explained here:

https://math.stackexchange.com/q/2521219/

A special function call (C-DEBUG-BREAK-AT/COMPENSATE) is run when the
usermode PARSE-based command-line processing is finished.  It takes an
upper bound on how many ticks may or may not have happened in order
to process the `--breakpoint` option.  The tick count is jumped ahead
according to a formula that can account for a variation of up to 1000
stray ticks.