Fix molding abnormalities
When code was scanned like:

    [[
        a
    ]]

The indentation logic of R3-Alpha would mold this as:

    [[
            at
        ]]

Basically, it was indenting when there were new levels of block if
there were any lines.  The indentation level only needs to increase
if a block molding brings in newline elements itself.

This adjusts that and some other tweaks to make molding more sane.

Also includes proper enablement of special editing functions for the
windows console, so that it's possible to select text with the mouse
and so that inserting new characters feeds the other characters
ahead in the console.