Fix MOLD/ONLY behavior after previous commit
While the previous commit adjusted the newline and indenting behavior
to be more in-line with expectation, this pulled the rug out from under
the expectations of MOLD/ONLY, resulting in that output not having any
newlines at all.

This corrects that issue.  It also tightens the NEW-LINE command to
give an error if the user tries to put a marker on the tail of an
array, as that is not supported in Ren-C at this time (nor in R3-Alpha
nor Red).