Fix reversed path error, tidy SET-PATH! invariant in eval
An error message incorrectly reversed the arguments of the PATH! and
the thing that could not be found in the path.  This fixes that, and
also makes it clearer what part of the path evaluation state is a path
(the name "orig" was not as communicative as "any_path").

A historical problem which path evaluation had regarding reading the
thrown bit out of an uninitialized cell has a cleaner way to address
today, so this removes lengthy comments and enforces with clearer
and more enforceable debug build checks.