Allow CHECKSUM to take BINARY! for the key
For some reason, CHECKSUM was only allowing ANY-STRING! for the key.
This meant that someone with a BINARY! would have to run it through
a codec that guaranteed it would not become a wide string.

This changes the TLS code and checksum to work with plain BINARY!,
and restricts the input to STRING! (as TAG! or FILE! can be aliased
AS STRING! and it makes it less error-prone to accept fewer types).