API name + ergonomic improvements (REBYTE->char for UTF-8)
This gives the awkwardly named functions for extracting data from
strings (or building them) better names (e.g. `rebSpellingOf`).

The REBYTE type is used internally as a precaution to distinguish
between one-codepoint-per-char encodings (ASCII, Latin1) and UTF-8.
By not using the (usually) signed char for UTF-8, it calls to attention
the idea that you can't use `strlen()` and get a number of chars.

As helpful as that may seem, it creates a speedbump for C programmers
to have to cast in order to work with the data.  So this switches to
only obeying the convention inside the core sources--not in the RL_API.

Also adds APIs for translating Rebol FILE! to local paths and back
(the equivalent of TO-REBOL-FILE and TO-LOCAL-FILE)