Add `PANIC*` for a "raw panic" on a value's series
The PANIC primitive which was added would let you PANIC on a STRING!
or an ERROR!, but it would assume you wanted to dump the tick count
and halt the interpreter while delivering that as the message.  Hence
it would extract the UTF8 string from a STRING! and panic on *that*.

This adds `PANIC*` which accepts ANY-VALUE! and does no preprocessing
on it.  So `panic* some-string` will dump out the tick count where the
SOME-STRING was created, and the relevant stack of creation if using
valgrind or address sanitizer.

Also fixes a problem where PANIC was quoting its argument when it
should have been evaluated normally.