Append_Codepoint_Raw and Temp_Bin_Str_Managed renamings
The routine Append_Codepoint_Raw() was given its name in order to warn
that it did not widen byte-sized strings if given a high codepoint.
This concept is going away in UTF8-Everywhere, so this turns it into
just Append_Codepoint()

Temp_Bin_Str_Managed has always been a confusingly-named routine, but
its job was to detect if a string was all ASCII and BYTE_SIZE, and if
so note that it could be used directly as UTF8...and if not, that a
temporary series would have to be created.  This is a routine that will
not be needed after UTF8-Everywhere.  For clarity in the meantime,
this renames it.

Both of these renamings are broken out into a separate commit so the
simple change of a rename won't cloud the readability of other changes.

Also switches two incorrect usages of VAL_BYTE_SIZE() to assume
correct UTF8 data in a string the crypto and net code, and defers to
the Make_Series_Codepoint() function instead of trying to create
a byte-size series directly in scanner errors.