Temporarily change convention for compression in LOAD-EXTENSION
In order to do "more of the work" in usermode, the C initialization
code for extensions would initialize a BINARY! if a script's data
was compressed, or a byte-sized STRING! if it was not.  This cued
LOAD-EXTENSION into knowing--based on the type--if it needed to call
DECOMPRESS or not.

Prior to UTF-8 Everywhere, there is to be a phase of "Latin-8 Nowhere",
which means no STRING! will be allowed to be byte-size.  Hence, this
trick won't work as the UTF-8 data for the uncompressed script is in
an unsuitable format to be stored directly into a STRING!.  And
it's not a good idea to expand it into wide character data, because the
scanner would then have to convert it back into UTF-8 again.

So this just wedges the decompression into the C code for extensions.
However, once UTF-8 only is implemented, this can go back to the same
trick (which means extensions won't be tied to using the internal
API that exposes things like Inflate_To_Series, but can just use the
RL_API)