Fix bug in unicode variation of Detab_Unicode
At some point a change in Ren-C moved an initialization of a variable
`n` up to its point of declaration, without realizing it was meant
to be initialized to 0 at the specific point it was.  This meant
the unicode version of Detab_Unicode gave an incorrect result when
removing the first tab.

Also switches the naming convention to the `up` for "unicode pointer"
(vs. "byte pointer") for the routine.