Specialize popping of BINARY! from mold buffer
Today's mold buffer is REBUNI-sized, but with UTF-8 everywhere it will
be a byte buffer.  Hence popping a binary will be the same as popping
a UTF-8 string.  However, until then it needs special case code, as
there will not be a "slimming" form of popping mold in Latin1-Nowhere.