TO-LOCAL-FILE/TO-REBOL-FILE as LOCAL-TO-FILE/FILE-TO-LOCAL
Rebol wants to standardize FILE! literals to speak in the forward-slash
format.  So `%\foo\bar` gets turned into `%/foo/bar`.

(Arguably, it should complain about the literal form and refuse to
LOAD it instead of losing information)

This tightens the data types such that LOCAL-TO-FILE takes a STRING!
and returns a FILE!, and FILE-TO-LOCAL takes a FILE! and returns a
STRING!.  Another feature added is the /ONLY flag, which permits
pass-thru of an already converted quantity.  So if LOCAL-TO-FILE/ONLY
gets a FILE! as input, it won't error but will return a copy of that
FILE! instead.

(Note: It would be possible to subvert and get re-conversion on a FILE!
by saying AS STRING! MY-FILE, but that would defeat the purpose of
using the datatype to distinguish converted vs. non-converted values.)