Fix buffer overrun when delivering scan errors on PATH!
Incomplete paths (such as `load {=/}`) were causing a potential
buffer overrun when delivering the error message.  The problem was that
there are local variables in Scan_Array() which--for convenience--are
not synced with the SCAN_STATE versions of those variables.  These
track the start and end positions in the input that are tokenized.
If the local variable for the end is updated without updating the
SCAN_STATE variable, and the SCAN_STATE variable for the beginning is
updated, it could wind up with the beginning seeming after the end
in the `Error_Syntax(ss)` which only takes the scan state.

This addresses the problem for the moment, though having the
information repeated for the sake of typing/readability may not be
worthwhile.  Calling the variables `ss->b` and `ss->e` isn't really
that much worse than `bp` and `ep`, and would be easy enough to read
for someone familiar with the code.