Don't use memcpy() to fill in parent context values
It is no longer legal to bit-copy arbitrary values from one cell to
another without masking out the "formatting" bits of that cell.  This
masking is done by Move_Value() and Derelativize().

Some code inside of constructing contexts that "inherit" from other
contexts was written to use a raw memcpy() and hadn't been revisited.
This changes it to a Move_Value() loop.