Replace OS_CRASH with HIJACK-able rebPanic() API
This makes rebPanic() an API which if it receives a UTF-8 string will
dispatch it to the PANIC native...and if it receives a REBVAL* value
pointer it will dispatch it to the PANIC-VALUE native.

By default, those natives do the debug-aware procedure of dumping as
much information as possible--the same as panic() when called from the
C core.  However, making them functions offers two advantages:

* Users can call PANIC or PANIC-VALUE directly from Rebol code, and get
  any information about them a debug build happens to have.

* A given Rebol host (such as the current HOST-CONSOLE) is able to use
  HIJACK to customize the reaction to an API-level panic.  This might
  be writing to a log or taking some kind of mitigating action, but
  not crashing.

This hooking only happens with libRebol's API-facing rebPanic(), not
with panic()...as the latter case ostensibly indicates a serious "blue
screen of death" in the internals of the core.  However, the core
is able to use this function as well.

Because of the flexibility of this HIJACK-able approach, it obsoletes
the need for the kludgey OS_CRASH() API, and %host-error.c.  That
meant eliminating the last bit of I/O done from the C main() program
itself vs. delegating to HOST-CONSOLE...now all messages about
"Ctrl-C pressed" or error reports are done in usermode.

So Put_Str() and several related C kludges could be killed as well.
It also goes ahead and gets rid of the backtrace buffers for C string
output--which is an interesting concept for the debugger to pursue
(the ability to look back in the evaluator history at past evaluations
without dumping them in realtime).  This concept will be explored in
usermode-based debug and trace, but C code for it is not prudent to
maintain.