ENCLOSE call to "inner" function in an "outer" function
This adds a new function generator called ENCLOSE.  It is similar to
a combination of ADAPT and CHAIN, but is strictly more powerful than
both (alone or combined).

ENCLOSE takes an outer function and an inner function.  The outer
function should be single-arity, and accept a fulfilled frame for
invoking the inner function (which it does with DO, as DO on a FRAME!
knows what function that frame is for).

This allows code to be run both before and after the wrapped function's
execution--as well as mutations to the wrapped function's frame before
running (similar to ADAPT).  But it can also carry state across the
execution and use it to mutate the result (whereas CHAIN has to be
a pure function of the output of the previous function in the chain)

Imagine you want a hooked version of multiply that always adds the
difference of the arguments to the result:

    >> e-multiply: enclose 'multiply function [f [frame!]] [
        diff: abs (f/value1 - f/value2)
        result: do f
        return result + diff
    ]

    >> e-multiply 7 10
    == 73

Here's mutating the frame before invocation, and choosing not to pick
a moment to run the DO at all:

    >> n-print: enclose 'print procedure [f [frame!]] [
        if #"n" = first f/value [leave]
        f/value: reverse copy f/value
        do f
    ]

    >> n-print "fun with ENCLOSE"
    ESOLCNE htiw nuf

    >> n-print "not printing because starts with n"
    ;-- no result

Since ENCLOSE needs to take a single argument, it needs to be a
FUNCTION! that takes a FRAME!...vs the convenience of a BLOCK! like
ADAPT that receives bindings for the frame.  However, IN can be used to
smooth over the problem (but less efficiently than using ADAPT):

    >> n-print: enclose 'print procedure [f [frame!]] [
        do in f [
           if #"n" = first value [leave]
           value: reverse copy value
           do f
        ]
    ]