Tail-Call recursions via REDO primitive
REDO is a function that takes a running FRAME! (or a means of finding
a running FRAME!--such as a WORD! of a local variable, or FUNCTION!
representing a definitional return) and causes it to restart the
function with its current state--without creating another stack frame.

This allows one to program in a style similar to recursion, but
without running the risk of stack overflows.  Consider this recursive
function definition:

    foo: func [n /accum a] [
       case [
           not accum [a: 0]
           n = 0 [return a]
       ]
       foo/accum (n - 1) (a + 2)
    ]

For small initial values of n, such as `foo 1000`, this works fine and
gives back 2000.  But even just `foo 10000` causes a stack overflow.

Using REDO, the function can be implemented in a similar style but
with only one stack frame:

    foo: func [n /accum a] [
        case [
           not accum [a: 0 | accum: true]
           n = 0 [return a]
        ]
        frame: context-of 'n
        n: n - 1
        a: a + 2
        redo frame
    ]

Even if you do `foo 1000000`, only a single stack frame will be used.
Each REDO will leave the arguments and refinements as they were, but
locals are cleared and types are re-checked.

The cleanest way to express a redo is probably to REDO :RETURN the
definitional return for a function.  It could also be a refinement
to the definitional return e.g. RETURN/REDO, however since RETURN
takes an argument to return that would be necessarily discarded.
This makes REDO :RETURN a fairly decent idiom.

    foo: func [n /accum a] [
        case [
           not accum [a: 0 | accum: true]
           n = 0 [return a]
        ]
        n: n - 1
        a: a + 2
        redo :return
    ]