Use c-style cast in SYM_FUNC()
Something about C++14 gcc and some clang doesn't like the cast()
macro for CFUNC*.  It also more aggressively caught a missing const
on a cast from `const void*` to `char*`, which should have been
caught otherwise.  This is seen in release builds, where the cast
helpers are employed (since they are optimized out).

Fix the missing const and switch to C-style cast for the moment in
SYM_FUNC(), with a note to come back to the issue later.

Also fixes inconsistent line endings in %redo.test.reb