Allow "Sibling Tail Calls" via REDO/OTHER
Ren-C has a sophisticated model of frame reuse in function composition.
Functions like ADAPT and SPECIALIZE are able to reuse a frame, as
opposed to adding layers to the call stack.  This makes them similar
in nature to tail-calls.

The original conception of REDO was that the only functions you could
REDO would be ones that were part of the current composition.  By
definition, those functions all have a compatible frame layout.  Yet
it's also possible for frames to be compatible when they are not
part of the same composition.  e.g. two distinct adaptations of the
same underlying function both have frames compatible with the
underlying function, hence also with each other.

The idea of doing a tail call optimization which jumps to the top of
another function is called a "sibling tail call".  This commit lets
one do such calls for any frame-compatible FUNCTION! with the frame
(or implied frame) from the first parameter to REDO.

One mechanical aspect of making this work (which was an oversight in
the previous REDO commit) is that if the jump is to a function with
specializations in an exemplar frame, those specializations must be
reset.  Essentially, a REDO or REDO/OTHER should not be able to do
anything a recursion could not--that means no injection of locals, and
no bypassing expected specialization state.

As an additional change in this commit, the REDO :RETURN idiom is
deprecated in favor of REDO 'RETURN, due to issues of efficiency.
(:RETURN must run a GET to fetch the FUNCTION! and then read the
binding property, whereas there is no real evaluation of a LIT-WORD!
'RETURN beyond changing its type in the header, and its binding can
be extracted directly.