Use frame output cell vs. frame eval cell for Do_Signals output
Each frame has one "scratch cell" which is used for many purposes,
and implicitly guarded by the GC.

One of the scratch purposes the cell was being used for was to receive
thrown values from the "Do_Signals()" processing loop.  This is a
case where--for instance--a keypress could spontaneously interrupt
evaluation and cause an interactive debug session.

However, there was an expectation that the scratch cell not be
changed during the signal processing...and it was in the case of
reifying a frame from a va_list when rebEval() was being used.

Because this is difficult to address in a general way, this changes
things to use the frame's *output* cell.  This makes more sense, as
the result of the Do_Signals would be thrown and thus put into the
output cell.  But also, it lets the frame's cell be owned by the
evaluation process for whatever invariants that needs, vs. overwriting
it randomly.

Yet the output cell couldn't be claimed immediately, because it was
carrying the data of the previous result for enfix processing, even
when Do_Signals() was being called.  This suggests avoiding calls to
Do_Signals() until an enfix is finished (although a deeper frame can
use its cell to do the signals).

This also corrects a problem in claiming where an "expression starts"
in cases of enfix.  It should claim the entire expression now, e.g.
`1 != 2`, instead of `!= 2`.  (This is only noticed in error messages
today, but will become relevant in debugging.)

The invariant for the output cell in the expression loop is tightened
here to be forced to an unreadable blank.  This should help catch
accidental leaks of influence from previous evaluations (except in
the cases where the leak is intended and handled specially, e.g. enfix)