LENGTH/HEAD/CONTEXT/... via REFLECT, enfix REFLECT as OF
This commit moves more properties in to use the REFLECT mechanic.
They include LENGTH, INDEX, HEAD, TAIL, CONTEXT, etc.  This is combined
with an enfixed form of REFLECT, called OF...which makes a generic
way of accessing the reflector without needing to specialize it.

While it is currently still possible to access the properties in their
historical way, a new way opens up that does not require a hyphen, and
does not take away a variable name:

    print 1 + length of foo
    head: head of bar

Empirically this is pleasant to use, and appears less "warty" than
having to use the hyphens:

    print 1 + length-of foo
    head: head-of bar

And it's not as semantically problematic as the Rebol2/R3-Alpha way:

    print 1 + length? foo
    head: head bar

(LENGTH? is misleading because it returns an INTEGER!, not a LOGIC!.
HEAD is noun-like and a common variable name, and the above would
overwrite the HEAD operation so it could not be used again.)

Existing hyphenated specializations, and the classic HEAD and TAIL,
are kept for now...making this a choice and not an enforcement.  Many
questions remain about how this can be generalized, and those questions
fold into existing questions about user-defined types...how people
can add more actions on a type, or a type to use existing actions.

This commit is relatively invasive to unkempt PORT! and GOB! code.  The
updates try and make the change while sending a wave of improvement to
help more clearly see what needs to be done in the mechanics of
ports/actions/reflectors.