Fix edge case of quoting enfix after an evaluation
When judgments were being made about enfix after an evaluation, it
would look at a WORD! in that position and test to see if it looked
up to an enfix function.  If it did not, it would jump to the same
code that would have handled the REB_WORD anyway...reusing the lookup
as a tool for efficiency.

This did not account for the case where a non-enfix word might sit
after an evaluation, but be expected to be backwards quoted by another
operation after it.  To handle that case, it must jump up to the
reevaluate label (which also will be able to reuse the work from the
fetch).