Examine all parameters for function cached flags
This is a small preparatory change for some of the work on COMMENT and
other functions.  It visits all a functions parameters for the cache
step, gets rid of the `---` form of COMMENT (also not present in Red
or Rebol2), corrects some small typos and fixes FFI compilation.

Also adds DO_FLAG_FULFILLING_SET to SET-WORD! and SET-PATH! midframes.

Fixes small bug in DO of a VARARGS! at END