Invisible mechanic: COMMENT mostly invisible, ELIDE fully so
This introduces two new categories of "invisible" functions.  They are
designed to return nothing--not even void.

The way to say how a function is in this category in the moment is to
give it a `return: []` in its spec, to say the null set (even excluding
`<opt>` for void) is what it can return.  This means it will do its
best to omit any influence from the chain of evaluator (outside running
the body of the function itself).

Category #1 is when the function is not enfix.  This simplification
will attempt to complete the left-hand side of an expression before
running.  This gives a mostly-intuitive order of evaluation, as shown
by the DUMP debug primitive--which has been changed to be invisible:

    >> x: 10
    >> y: 20
    >> z: 1 + 2 * 3 dump [z y x]

    z: => 9
    y: => 20
    x: => 10
    == 9

The cost of such a model is that it can't be truly invisible, since
it has intruded on what the evaluative order would have been as seen
by any enfix operations on their right.  e.g. forcing (print 1) to
complete below means the 1 is no longer available for the (1 + 2)
that would happen if the comment weren't there:

    >> print 1 comment "hi" + 2
    ** Script Error: + requires value1 argument to not be void

Category #2 is achieved by making such a function enfix.  By convention
it will retain the state of the frame's output cell, instead of
erasing it after extracting into the left hand parameter slot of the
function.  This gives that function access to the state being passed
through it--as well as allowing the evaluator to retain enough of its
internal state to make it "fully invisible" regarding characteristics
not exposed (such as: there is no END! datatype to pass through to
write an `<end> status on the output).

A basic operator which provides commenting and evaluation in this
model is provided called ELIDE.  It can be used to erase any code if
passed a BLOCK!

    x: 10
    y: 1 elide [+ 2
    z: 30] + 7

It can also be passed a GROUP!, but the order of evaluation may be
surprising due to its tight-enfixing character.  For instance, the
ELIDE below must be done at the moment of evaluation of the 2, not
deferred after `y: 1 + 2` has all been completed:

    >> y: 1 + 2 elide (print y)
    ** Script error: y has no value

(Note: This commit includes clarifying code rearrangements which
involved changing indentation levels, so it appears larger than it is.)