Change `|>` and `<|` implementation, enfix ALSO (temp ALSO-DO)
This changes the "left bar" and "right bar" operators to take advantage
of the "eliding" nature of invisible functions from the previous
commit.  Before they had to be written as enfix, which limited the
amount of completion they could do on the left.  This reimagines them
to complete the left hand expression.  Hence:

    >> x: 1 + 2 * 3 <| print x y: x + 1 print y
    9
    10
    == 9

It also implements the proposal to change ALSO to an enfix function.
But to make it easier for people to find out about the change, the
ALSO function gives an informative error and suggests temporarily
using ALSO-DO (or enfixing the prefix form, AFTER, in one's own code)