Fix DO_FLAG_EXPLICIT_EVALUATE and VALUE_FLAG_EVAL_FLIPs
Do_Core() can be in a mode in which all values are evaluative, unless
they are passed as arguments to a function which quotes them.  (This is
the default.)  It can also be in a mode where no values are evaluative,
and are passed as-is to any argument slot.

To make the second mode at all useful, there is a bit that can be
flipped on any particular value to make it evaluator-active.  This same
bit also flips the evaluator in the default mode so that a value will
act quoted, even if it's being passed to an evaluative function
parameter.

This mechanism was invented for the libRebol API, which wants REBVAL*
passed as variadic arguments in C to be considered "already evaluated".
It is new, so the full implications of the bit flip are not completely
understood.  However, in the #tight parameter case, it wasn't passing
on the evaluative flag to the nested frame fulfilling that parameter.

There was another case where the right hand side of a SET-PATH! or a
SET-WORD! where it was not accounting for the possibility of the
bit being flipped, only the state of the overall evaluation.

Includes some reorganization for readability in addition to the fix,
and properly groups/indents some the post-apply checking so it's all
under the debug build only.  Also makes it possible to pass in
further flags to the Do_Next_Mid_Frame_Throws().  (Doesn't change
anything here, but useful for a later commit and clearer to see here
as part of a more minor change.)