Do_Core() sets f->out to END on each entry
Early incarnations of the Ren-C REBFRM* based evaluator attempted to
optimize to avoid recursions of Do_Core.  This meant it was sometimes
elided and re-entered in a special mode for certain enfix operations.

As the logic behind those operations became more complex, it made less
sense to avoid the recursion.  Instead, effort shifted to trying to
reuse the memory for the frame.  (See Do_Next_Mid_Frame_Throws())

Those early incarnations had conditional re-entry for optimized
enfix in which the f->out cell would not be END.  The invariant became
somewhat confusing, with a lot of redundant SET_END() calls and
hedging in the asserts.

This makes Do_Core()'s relationship to f->out (slightly) clearer, and
updates some asserts/errors related to recent changes for "invisible"
functions.