Consolidate Do_Core_Entry_Checks into Push_Frame
Given the length of Do_Core(), keeping the debug-build-only code down
to an informative minimum required separating out some of the heavier
invariant checking into routines in %d-eval.c.

There were checks on entry, checks on expression, and checks on exit.
But since an expression check is done immediately after entry, it
didn't really make sense to check very much unless it was checking
properties of the frame itself--some of which was second-guessing
setup in Push_Frame(), as well as doing checks that Push_Frame()
could do for all applications.

The checks are now moved to Push_Frame, which is a better place.
(Exception may be checks on f->out's suitability as a target, since
the same frame might use several different f->out slots.  This
suggests maybe a SET_FRM_OUT() function is needed.)